% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{jenkins}
\alias{jenkins}
\title{Jenkins}
\usage{
jenkins(server = "http://jenkins.ropensci.org", username = "jeroen",
  token = jenkins_pat(), verbose = FALSE)
}
\arguments{
\item{server}{naked url of the server}

\item{username}{name of the user to login}

\item{token}{authentication token (or password)}

\item{verbose}{show http output}
}
\description{
Some wrappers to control jobs and builds on a Jenkins server.
}
