% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{jenkins}
\alias{jenkins}
\title{Jenkins}
\usage{
jenkins(server = "http://jenkins.ropensci.org", username = "jeroen",
  token = jenkins_pat(), verbose = FALSE)
}
\arguments{
\item{server}{base url of the jenkins server}

\item{username}{name of the jenkins user to login}

\item{token}{authentication token (or password) for your jenkins server.}

\item{verbose}{print http output for debugging}
}
\description{
Managing jobs and builds on a Jenkins server. You can set a global Jenkins
access token via the \code{JENKINS_PAT} environment variable.
}
\section{Methods}{

\Sexpr[results=verbatim, stage=build, echo=FALSE]{jenkins::jenkins('dummy')}
}

\examples{
\donttest{# Make a connection
jk <- jenkins(server = 'http://jenkins.ropensci.org', username = 'jeroen')

# Do stuff
jk$info()
jk$job_build('magick')
jk$job_status('magick')
}

}
\references{
\url{https://wiki.jenkins.io/display/JENKINS/Terminology}
}
