% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{jenkins}
\alias{jenkins}
\alias{config_template}
\alias{sync_jenkins_ropensci}
\title{Jenkins}
\usage{
jenkins(server = "http://jenkins.ropensci.org", username = "jeroen",
  token = jenkins_pat(), verbose = FALSE)

config_template(git_url)

sync_jenkins_ropensci()
}
\arguments{
\item{server}{naked url of the server}

\item{username}{name of the user to login}

\item{token}{authentication token (or password)}

\item{verbose}{show http output}

\item{git_url}{HTTPS git url of the target repository}
}
\description{
Some wrappers to control jobs and builds on a Jenkins server.
}
